section1 auto

+++ PROS/1010X_5/include/util.h
@@ -13,5 +13,5 @@ void motorset_drive(int Lpower, int Rpower);
 
 float gyro_read(Gyro port, int multiplier);
 
-int pe_batt(int analogport);
+float pe_batt(int analogport);
 #endif
